Refactored package names
Also replaced health check with HTTP status
check for TestContainers
