+++ testing/internals/worker.go
@@ -4,10 +4,13 @@ import (
 	"bytes"
 	"encoding/json"
 	"fmt"
-	. "github.com/CopilotIQ/opa-tests/gentests"
+	. "github.com/CopilotIQ/opa-tests/testing"
 	"io"
+	"math"
 	"net/http"
+	"runtime"
 	"strings"
+	"sync"
 )
 
 const (
@@ -68,4 +71,48 @@ func GetResult(r io.Reader) (bool, error) {
 		return false, fmt.Errorf("%v is not a valid bool", result)
 	}
 	return b, nil
+}
+
+func EstimateWorkers() uint {
+	cores := runtime.NumCPU()
+	if cores > 3 {
+		Log.Debug("running with 70%% CPU load on %d cores", cores)
+		return uint(math.Ceil(0.7 * float64(cores)))
+	}
+	return 1
+}
+
+func RunTests(tests []TestUnit, workers uint, addr string) *TestReport {
+	dataChan := make(chan TestUnit)
+	var wg sync.WaitGroup
+	if workers == 0 {
+		workers = EstimateWorkers()
+	}
+	if workers > 1 {
+		Log.Info("running %d parallel test runners", workers)
+	} else {
+		Log.Warn("running single-core, execution will be slower")
+	}
+	url := fmt.Sprintf("http://%s", addr)
+	var report TestReport
+	for i := uint(0); i < workers; i++ {
+		wg.Add(1)
+		go func(num uint) {
+			Log.Debug("starting worker #%d", num)
+			err := SendData(url, dataChan, &report)
+			if err != nil {
+				Log.Error("error sending requests to OPA server: %v", err)
+			}
+			wg.Done()
+			Log.Debug("worker #%d done", num)
+		}(i)
+	}
+	for _, req := range tests {
+		dataChan <- req
+	}
+	// Once you're done sending data, close the channel
+	close(dataChan)
+	wg.Wait()
+	fmt.Println()
+	return &report
 }
